update windows installer

+++ linux/conda_quick_install.sh
@@ -2,11 +2,16 @@
 
 # 创建目录并下载 Miniconda 安装脚本
 mkdir -p ~/miniconda3
-wget --tries=10 --timeout=60 --waitretry=5 --no-check-certificate https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh
+if ! wget --tries=10 --timeout=60 --waitretry=5 --no-check-certificate https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O ~/miniconda3/miniconda.sh; then
+    echo "downloading miniconda.sh failed."
+    exit 1
+fi
 
 # 安装 Miniconda
-bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3
-
+if ! bash ~/miniconda3/miniconda.sh -b -u -p ~/miniconda3; then
+    echo "Miniconda 安装失败。"
+    exit 1
+fi
 # 删除安装脚本
 rm -rf ~/miniconda3/miniconda.sh
 
@@ -14,5 +19,4 @@ rm -rf ~/miniconda3/miniconda.sh
 ~/miniconda3/bin/conda init bash
 ~/miniconda3/bin/conda init zsh
 
-# 重新加载 shell 以应用 conda 初始化
-# exec $SHELL
+exit 0
